Fix threading issues
This ensures that if VST windows are open, that everything related to
those windows is done in the main UI thread.  Additionally, this ensures
that the shut down is in the main UI thread by invoking closeEditor to
the UI thread in the vst_destroy function, then calling deleteLater to
automatically delete itself in the UI thread after that action has been
completed.