Fix mvMeter2 resize crash, fix SPAN resizing
While `audioMasterSizeWindow: return 0` seems to work for some, it's
actually supposed to return 1. Because of this, Voxengo SPAN would never
resize its QWidget, and TBProAudio mvMeter2 would have unlocked sizing -
allowing it to resize the Qt EditorWidget into the negatives, causing
Qt / Windows to panic & crash.

With the above change locking down VST-related quirks, this left another
issue - resizing above the initial size would resize the Qt EditorWidget
but not the child QWidget, resulting in both the VST and the EditorWidget
having the correct size but the QWidget keeping its default size.
This is apparently a quirk in Qt specific to createWindowContainer. The
apparent solution (and the one used here) is to set the container's
parent to a nullptr and manually set its parent via a layout.

Using either one of these solutions without the other would introduce
other quirks in SPAN and Blue Cat Audio's VST, such as resizing *down*
causing the QWindow to incorrectly calculate its size and cropping off
parts of the VST while both the VST itself and the EditorWidget had the
correct size.